pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "phishing-surface-monitor"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pydantic>=2.0", "click>=8.0", "rich>=13.0", "dnspython>=2.4"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[tool.pytest.ini_options]
addopts = "--cov=analyzers --cov=schemas --cov-report=term-missing --cov-fail-under=70"
